requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aniso8601==9.0.1
    # via
    #   -r requirements.in
    #   flask-restful
apscheduler==3.8.0
    # via
    #   -r requirements.in
    #   flask-apscheduler
certifi==2022.6.15
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==2.1.0
    # via
    #   -r requirements.in
    #   requests
click==8.0.3
    # via
    #   -r requirements.in
    #   flask
colorama==0.4.4
    # via
    #   -r requirements.in
    #   click
datetime==4.3
    # via -r requirements.in
flask==1.1.2
    # via
    #   -r requirements.in
    #   flask-apscheduler
    #   flask-caching
    #   flask-cors
    #   flask-jwt-extended
    #   flask-restful
    #   flask-sqlalchemy
flask-apscheduler==1.12.2
    # via -r requirements.in
flask-caching==1.9.0
    # via -r requirements.in
flask-cors==3.0.9
    # via -r requirements.in
flask-jwt-extended==4.3.1
    # via -r requirements.in
flask-restful==0.3.8
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via -r requirements.in
greenlet==1.1.2
    # via
    #   -r requirements.in
    #   sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via
    #   -r requirements.in
    #   requests
itsdangerous==2.0.1
    # via
    #   -r requirements.in
    #   flask
jinja2==3.0.1
    # via
    #   -r requirements.in
    #   flask
markupsafe==2.0.1
    # via
    #   -r requirements.in
    #   jinja2
pyjwt==2.0.0
    # via
    #   -r requirements.in
    #   flask-jwt-extended
pymysql==1.0.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   flask-apscheduler
pytz==2021.3
    # via
    #   -r requirements.in
    #   apscheduler
    #   datetime
    #   flask-restful
    #   tzlocal
requests==2.28.1
    # via -r requirements.in
shortuuid==1.0.1
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   apscheduler
    #   flask-cors
    #   flask-restful
    #   python-dateutil
sqlalchemy==1.4.31
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
tzlocal==2.1
    # via
    #   -r requirements.in
    #   apscheduler
urllib3==1.26.11
    # via
    #   -r requirements.in
    #   requests
uuid==1.30
    # via -r requirements.in
werkzeug==2.0.2
    # via
    #   -r requirements.in
    #   flask
    #   flask-jwt-extended
zope-interface==5.4.0
    # via
    #   -r requirements.in
    #   datetime

# The following packages are considered to be unsafe in a requirements file:
# setuptools
